refactor: PopularFeedsEntity 제거 및 PopularFeedEntity 리스트 사용 방식으로 변경
- `PopularFeedsEntity.kt`: 불필요한 래퍼 클래스인 `PopularFeedsEntity`를 삭제하고, 내부 `PopularFeedEntity`를 별도 파일로 분리
- `FeedMapper.kt`: `toData` 함수의 반환 타입을 `PopularFeedsEntity`에서 `List<PopularFeedEntity>`로 수정
- `FeedRepository.kt`: `fetchPopularFeeds`의 반환 타입을 리스트로 변경하고, 필터링 로직을 리스트 확장 함수를 사용하는 방식으로 간소화
- `HomeViewModel.kt`: 래퍼 클래스 제거에 따라 `popularFeeds` 데이터 접근 로직 수정
- `HomeUiState.kt`, `PopularFeedsAdapter.kt`, `PopularFeedsViewHolder.kt`: 참조하던 `PopularFeedsEntity.PopularFeedEntity` 타입을 `PopularFeedEntity`로 업데이트
