+++ app/src/main/java/com/into/websoso/data/model/PopularFeedEntity.kt
@@ -1,0 +1,13 @@
+package com.into.websoso.data.model
+
+data class PopularFeedEntity(
+    val feedId: Long,
+    val feesContent: String,
+    val likeCount: Int,
+    val commentCount: Int,
+    val isSpoiler: Boolean,
+    val isPublic: Boolean,
+    val novelTitle: String,
+    val novelImage: String,
+    val novelGenre: String,
+)
